aamachine-0.4.4
0.4.4:

C64 interpreter: Added French accented characters to the font.

Web interpreter: Improved support for scrolling with page up
and page down when the input field is active. This failed on
some browsers before.

aambundle: Added target "web:story" for generating just the
story.js file for the web interpreter. This can sometimes make
it easier to fit the Å-machine tools into a larger build
process. Be aware that the web interpreter also expects to find
the original .aastory file in the resources directory, with a
mangled filename.